Split sample browser hit target widget
